Apply changes from code browser
Apply changes from code browser
